+++ src/util/EntityLoaderUtils.cpp
@@ -11,7 +11,7 @@ int EntityLoaderUtils::GetTextWidth(std::string text, int height)
 	sf::Font font;
 	font = *FontContainer::GetFont(ApplicationParameters::k_fontPath);
 	dummyText.setFont(font);
-	dummyText.setCharacterSize(height*ApplicationParameters::k_textFactor);
+	dummyText.setCharacterSize(static_cast<float>(height*ApplicationParameters::k_textFactor));
 	dummyText.setString(text);
-	return dummyText.getLocalBounds().width;
+	return static_cast<int>(dummyText.getLocalBounds().width);
 }
